Adds a new folder for processing conversions
This folder serves as a location to upload code conversions from Adam's
iControl application to the remote folder within mousetrap for group-wide
development without interfering with the rest of the mousetrap
application.

Included are conversions of the Point and PointBuffer classes from Adam's
code.  Limited testing has been performed as the code base for the
application is incomplete.